Moved all restframework code to the rest_framework folder
