+++ api/app/signals/apps/questionnaires/rest_framework/views/__init__.py
@@ -1,10 +1,10 @@
 # SPDX-License-Identifier: MPL-2.0
 # Copyright (C) 2021 Gemeente Amsterdam
-from signals.apps.questionnaires.views.private import (
+from signals.apps.questionnaires.rest_framework.views.private import (
     PrivateQuestionnaireViewSet,
     PrivateQuestionViewSet
 )
-from signals.apps.questionnaires.views.public import (
+from signals.apps.questionnaires.rest_framework.views.public import (
     PublicQuestionnaireViewSet,
     PublicQuestionViewSet,
     PublicSessionViewSet
