+++ api/app/signals/apps/questionnaires/tests/rest_framework/views/__init__.py
@@ -1,0 +1,2 @@
+# SPDX-License-Identifier: MPL-2.0
+# Copyright (C) 2021 Gemeente Amsterdam
