+++ api/app/signals/apps/questionnaires/tests/rest_framework/views/test_private_questionnaires_endpoint.py
@@ -39,10 +39,10 @@ class TestPrivateQuestionnaireEndpoint(ValidateJsonSchemaMixin, APITestCase):
         self.questionnaire = QuestionnaireFactory.create()
 
         self.detail_schema = self.load_json_schema(
-            os.path.join(THIS_DIR, '../json_schema/private_get_questionnaire_detail.json')
+            os.path.join(THIS_DIR, '../../json_schema/private_get_questionnaire_detail.json')
         )
         self.list_schema = self.load_json_schema(
-            os.path.join(THIS_DIR, '../json_schema/private_get_questionnaire_list.json')
+            os.path.join(THIS_DIR, '../../json_schema/private_get_questionnaire_list.json')
         )
 
     def test_questionnaire_list(self):
